finance transaction created for monthly payslip after approval via delayed job, fixes #3316
git-svn-id: file:///svn/fedena2.1/branches/fedena2.1_i18n@3433 1b8ff446-76f8-11e0-a33c-ba5893e5f458

+++ app/controllers/employee_controller.rb
@@ -1232,13 +1232,13 @@ def one_click_payslip_revert
       payslip_record = MonthlyPayslip.find_all_by_employee_id(e.id,
         :conditions => ["salary_date >= ? and salary_date < ?", start_date, end_date])
       payslip_record.each do |pr|
-        pr.destroy
+        pr.destroy unless pr.is_approved
       end
       individual_payslip_record = IndividualPayslipCategory.find_all_by_employee_id(e.id,
         :conditions => ["salary_date >= ? and salary_date < ?", start_date, end_date])
       unless individual_payslip_record.nil?
         individual_payslip_record.each do|ipr|
-          ipr.destroy
+          ipr.destroy unless ipr.is_approved
         end
       end
     end
